Change eventNum from numeric to integer

+++ R/as_das_df.R
@@ -44,7 +44,7 @@ as_das_df.data.frame <- function(x) {
     Data7 = "character",
     Data8 = "character",
     Data9 = "character",
-    EventNum = "numeric",
+    EventNum = "integer",
     file_das = "character",
     line_num = "integer"
   )
